feat: [IXSPD1-2730] show apr in pool detail
